Rework gitignore to function as allowlist
Reworking of .gitignore files to exclude everything up front, with
exceptions for files and folder that should be included. This is to
reduce the chance of pushing any files unintentionally.

As part of this rework, a `schema/revert/templates` folder and files were found
and should not be there. They already exist in `schema/templates`
where they belong, so are being dropped from `schema/revert`.