Re-Add `added_surveillance_groups` to the Priority Queue
Last fall, the `added_surveillance_groups` CTE was removed from the priority
queue. This invite logic is once again desired, so this change adds it back
in. Tested locally by adding a value for this field to encounters and making
sure they got pulled into the queue at the proper priority.