Modify clinical ETL to process PHSKC data into FHIR bundles
Updates to clinical ETL for processing of PHSKC data. This data has being parsed and pre-processing
from Excel worksheets, and inserted into the clinical receiving table. This update to the clinical
ETL will transform the data into FHIR bundle format and inserted into the fhir receiving table to be
processed by the FHIR ETL.

This was modeled after the UW Retrospectives process, which is currently the only clinical data ETL that
uses the FHIR format. To expediate the development and avoid duplicating code, it utilizes existing
functions that were originally designed for REDCap ETLs wherever possible.