Add PHSKC `result_value` to FHIR Bundle under an Observation
This change mimics how the UW Retrospective ETL handles external test results and
includes functionality to add the `result_value` field for PHSKC data into Observation
resources consistent with the FHIR format. Slight modifications were made to the existing
functions used for this pipeline in the `fhir` module for readability.