Add API endpoint for deliverables log
Adding an API endpoint to retrieve records from the deliverables log table, for use
by LIMS for QC processes.

Since entries in this table only include sample OR collection barcode (depending on the
process being logged) and LIMS does not have a record of the sample identifiers assigned
to tiny swabs, this endpoint will return the associated sample and collection barcodes for
each log entry. The `details` returned for each record are limited to values that may be useful
for QC.

Access to this endpoint is being limited to localhost and LIMS IP address which are the only
current use cases.