Add shipping view for SCAN encounters with best available vaccination data.
Data quality issues were previously identified in the SCAN REDCap project vaccination
date and manufacturer fields, with participants sometimes entering invalid or conflicting
data during one or more encounters. To determine the most accurate vaccination status with
the available data, this shipping view is being added.

The logic being used here is first to identify the most-frequently reported valid vaccination date
for each dose. Valid dates are considered to be those between 12/01/2020 (based on the wording
of the questionnaire in REDCap) and the date of the encounter they were entered on (to exclude
future dates entered). The most frequently-reported (or only) valid date for each dose is then
applied to all encounters for that participant. The vaccination status for each encounter is then
calculated based on those applied dates and associated manufacturer values.