Include sort on encountered date when identifying most frequent vaccinnation date
Including an addition sort on encountered date (ascending) to more consistenly handle
cases where there is a tie on the number of times the same vaccination date was reported.
In these cases the earlier encounter date will be closest to the actual vaccination date,
which is assumed to be the most accurate.