Only Dump PHSKC Data if DF is non-Empty
PHSKC data is uploaded to sharepoint prior to the lab adding it to AQ sheets.
When this data is parsed, the outputted files are non-empty but with a new line,
causing issues when trying to check if there is data. This change only outputs
PHSKC data when the DF is non-empty, ensuring we are able to check files downstream
for emptiness.