New version of DOH linelist shipping view
A new version of DOH Linelist shipping view is being to added to include the following changes:

There are cases where records in REDCap are marked as not testable, but we receive
results from Samplify the referenced barcode. Adding an additional filter to ensure that
these samples (which have RoR result status of `never-tested`) are not included in DOH
linelist submissions.

For more accurate reporting, `date_tested` will use `result_ts` going forward and
`hcov19_result_release_date` for records prior to this change.

Date filters are being added to only include AFH/Workplace results after 2021-11-21 (when this
project was added to linelist process) and all other results after 2021-05-12 (when linelist
automation began).

Appropriate permissions are added to `return-results-exporter` role on the new view.