Bump UW Retros ETL Revision Number
As part of changes to the code structure of our ETL routine we are also
bumping the revision number of UW retro records so they are reingested.
When work first started on this data we used an old UK snomed Covid code,
which we can now replace with the international version. As a result, all
records ingested with the old code should be re-ingested so they reflect
this change.