Change Warehoused Covid Code from UK Snomed to International Snomed
This change alters the Snomed code associated with Covid-19 targets from the UK version
to the international version in our warehouse target table. The refactor of clinical
functions altered our coding associations to use this code instead of the UK version
to match our other coding schemes with Snomed codes.