Add check for more swab kits sent than triggered
To reliably associate each sample with a weekly or symptom instrument, the
number of kits sent each week should not exceed the number that were triggered.
If there were more kits sent than expected in a given week, that record is skipped
and a warning logged, after which the record needs to be fixed in REDCap.