Remove Old Snomed Code From Target Table Definition
We are now using a new Snomed code for Covid, under target 952. This change
removes the outdated target from the definition of the table. PA results
with this old target (789) were transitioned to the updated target (952).
When this sqitch change is deployed, we can also remove the outdated target
from the warehouse table and continue using 952 for future snomed cov targets.