Handle New Vaccine Date Format in UW Retros
We recently filled in some missing date in a handful of UW Retros records in REDCap,
but some date fields were populated with a non ISO date format (MM/DD/YYYY) which
causes issues with downstream fhir processes. This change adds a conditional that
reformats this date string into ISO format for compatibility with downstream processes.