Force PHSKC Barcode and CID Comparisons to be Uppercase
Original implementation of the PHSKC ETL assumed that barcodes and CID values
would always be provided as uppercase. This turned out to not be the case, and
this caused many samples from October 2022 to not be linked to the associated
metadata. This change forces all columns which will be compared for metadata
associations to be uppercase.