Add LIMS Identifier Matching Functions to Clinical Routines
Adds two functions that enable us match external clinical identifiers
with internal SFS identifiers stored in the LIMS.

`prepare_lims_request` takes various properties of our request to the LIMS
and builds an appropriately authenticated request, then prepares it. After
constructing a request with this function, it is ready to be sent to the
LIMS. You may pass credentials to this function, but if these are not provided
we fall back on environment variable values. If these are not found, this
function will raise an error.

`match_lims_identifiers` takes a dataframe of clinical records and a dictionary
of clinical/lims identifiers. For some routines like PHSKC, we are provided
multiple possible identifiers but the LIMS only stores these under one field.
Thus this dictionary should be formatted like `{clinical_term: lims_term, ...}`.
For our PHSKC example, this looks like the following:
```
{
  'main_cid': 'phskcCid',
  'all_cids': 'phskcCid',
  'phskc_barcode': 'phskcCid'
}
```

This allows us to search for multiple external identifiers in one go and construct
a dictionary of all possible matches. We'd even be able to search across multiple
LIMS identifiers in one request, so long as each identifier we are querying is
passed to this function. We return a dictionary where each identifier is matched
to the sample barcode associated to it in the LIMS. Note that non-matching identifiers
are excluded from this dictionary instead of being matched to `None`.