Refactor `parse-phskc` for Manifest. Add Deduplication Command
Refactors the `parse-phskc` clinical command to parse into a manifest
file. We now require a manifest filename and a file pattern in addition
to our geocoding cache file when invoking this function.

We first read in our manifest file to get a dataframe of all previously
parsed records. For each file matching the provided pattern, we get
all previously matched records from that file. If these records were
parsed more recently than the file was last updated, we don't reparse
them. If a file was more recently updated or there is no previously
parsed data from the file, we parse it and add it to our manifest data.

The parsing step remains the same except we add a `last_parsed` timestamp
to each parsed record. These will be dropped before being ingested by ID3C.
Note also that this command used to remove PII. This is no longer the case,
and this manifest file contains PII. It should remain on S3 only.

All parsed records (both newly and previously parsed) are output to stdout
after each command execution. You will likely want to redirect these records.
You can find any newly parsed records by running `id3c manifest diff <old parse>
<new parse>`. Since the last parsed timestamp is guaranteed to have changed,
these hashes will be updated and newly parsed records will show up in the diff.

This commit also adds an additional `deduplicate-phskc` clinical command. This
command takes in a manifest file with duplicate records (PHSKC sends us data
that contains duplicates across files) and removes those duplicates. For PHSKC,
there are three duplicate classes:
- Fully duplicated records: records that have all identifiers and metadata
duplicated. We keep the last record in the file, as this is the most recently
parsed of the duplicate records.
- Records with all identifiers duplicated: records that have all identifiers
duplicated but with non-duplicated metadata. We drop all these records, since
we don't know which metadata should be ingested.
- Records with one identifier duplicated: records that have one identifier
duplicated but with non-matching metadata. We drop all these records for the
same reason as above.

This deduplicate command will dump a version of the manifest that is passed in
to stdout, with duplicates removed.