Add `phskc-match` Clinical Command
Adds a new `match_phskc` clinical function. This takes in parsed data we would
like to match with LIMS identifiers. It also takes in a file containing previously
unmatched records and a file containing previously matched records. The goal is to
attempt to match any passed parsed data, as well as the records in the unmatched
record file.

We first loop through any files present in the provided parsed data and drop any
records from these files in our currently unmatched data. We prefer to match on
fresly parsed data rather than stale unmatched data. All records in this newly
expanded unmatched dataframe are then queried against the LIMS to try and match
them by `phskcCid`.

If any records match, they are added to the match data. Stale barcodes are dropped
from this dataset and freshly matched data is added. All matched data is then output
to stdout and the unmatched data file is updated based on what we were unable to
match on this run.

PII is removed from matched records, but not from unmatched records. We will continuously
try to rematch data on subsequent invocations of this command, assuming the file is
provided. By comparing the output of this command with the provided matched manifest
file, you can find which records were newly matched. These records can then be uploaded
to ID3C for ingestion via the existing clinical ETL routine.