Cast `census_tract` to `string` Instead of `int64`
When reworking the PHSKC ETL, it became clear that the census tract had
to be recast when it was read into a pandas dataframe. If it wasn't, it
would be cast as a float and data comparisons would fail during the
`manifest diff` step. However, this field should be cast as a string
instead of an integer so that assumptions about datatypes during downstream
ingestion via the FHIR ETL are not broken.